% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuniCandS-package.R
\docType{package}
\name{MuniCandS-package}
\alias{MuniCandS-package}
\title{MuniCandS: Multivariate Tests of Uniformity, Normality, Symmetry and Independence}
\description{
Implements multivariate tests of uniformity on the hypercube and hypersphere, as well as tests of normality, spherical symmetry, elliptical symmetry and independence. The procedures are of Cramér-von Mises type and are based on a decomposition of the empirical process into asymptotically independent components.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/emcabana/MuniCandS}
  \item Report bugs at \url{https://github.com/emcabana/MuniCandS/issues}
}

}
\author{
\strong{Maintainer}: Alejandra Cabana \email{AnaAlejandra.Cabana@uab.cat}

Authors:
\itemize{
  \item Enrique M. Cabana
}

}
\keyword{internal}
