% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MuniCandS-package.R
\docType{package}
\name{MuniCandS-package}
\alias{MuniCandS-package}
\title{MuniCandS: Multivariate Tests of Uniformity, Normality and Isotropy on C and S}
\description{
Implements tests of uniformity on the hypercube C_p = [0,1]^p and on the hypersphere S^p, and tests of normality and isotropy on R^p for p = 1, 2, .... The tests are of the Cramer-von Mises type, based on a decomposition of a p-parameter Brownian sheet as the sum of 2^p independent Gaussian processes.
}
\keyword{internal}
